refactor(idpe-17789): rename mod to compaction_job_done_sink
