+++ compactor/src/components/compaction_job_done_sink/error_kind.rs
